+++ test/foundry/SeaDrop-mintAllowList.t.sol
@@ -77,8 +77,8 @@ contract ERC721DropTest is TestHelper {
         MintParams memory mintParams = MintParams(
             0.1 ether, // mint price
             10, // max mints per wallet
-            uint64(block.timestamp), // start time
-            uint64(block.timestamp) + 1000, // end time
+            uint48(block.timestamp), // start time
+            uint48(block.timestamp) + 1000, // end time
             1,
             1000,
             100, // fee (1%)
@@ -132,8 +132,8 @@ contract ERC721DropTest is TestHelper {
         MintParams memory mintParams = MintParams(
             0.1 ether, // mint price
             10, // max mints per wallet
-            uint64(block.timestamp), // start time
-            uint64(block.timestamp) + 1000, // end time
+            uint48(block.timestamp), // start time
+            uint48(block.timestamp) + 1000, // end time
             1,
             1000,
             100, // fee (1%)
@@ -196,8 +196,8 @@ contract ERC721DropTest is TestHelper {
         MintParams memory mintParams = MintParams(
             0.1 ether, // mint price
             10, // max mints per wallet
-            uint64(block.timestamp), // start time
-            uint64(block.timestamp) + 1000, // end time
+            uint48(block.timestamp), // start time
+            uint48(block.timestamp) + 1000, // end time
             1,
             1000,
             100, // fee (1%)
@@ -265,8 +265,8 @@ contract ERC721DropTest is TestHelper {
         MintParams memory mintParams = MintParams(
             0.1 ether, // mint price
             10, // max mints per wallet
-            uint64(block.timestamp), // start time
-            uint64(block.timestamp) + 1000, // end time
+            uint48(block.timestamp), // start time
+            uint48(block.timestamp) + 1000, // end time
             1,
             1000,
             100, // fee (1%)
@@ -331,8 +331,8 @@ contract ERC721DropTest is TestHelper {
         MintParams memory mintParams = MintParams(
             0.1 ether, // mint price
             10, // max mints per wallet
-            uint64(block.timestamp), // start time
-            uint64(block.timestamp) + 1000, // end time
+            uint48(block.timestamp), // start time
+            uint48(block.timestamp) + 1000, // end time
             1,
             1000,
             100, // fee (1%)
@@ -397,8 +397,8 @@ contract ERC721DropTest is TestHelper {
         MintParams memory mintParams = MintParams(
             0.1 ether, // mint price
             10, // max mints per wallet
-            uint64(block.timestamp), // start time
-            uint64(block.timestamp) + 1000, // end time
+            uint48(block.timestamp), // start time
+            uint48(block.timestamp) + 1000, // end time
             1,
             1000,
             100, // fee (1%)
@@ -469,8 +469,8 @@ contract ERC721DropTest is TestHelper {
         MintParams memory mintParams = MintParams(
             0 ether, // mint price (free)
             10, // max mints per wallet
-            uint64(block.timestamp), // start time
-            uint64(block.timestamp) + 1000, // end time
+            uint48(block.timestamp), // start time
+            uint48(block.timestamp) + 1000, // end time
             1,
             1000,
             100, // fee (1%)
